[refactor]: Restructure map and login feature to MVVM architecture
