+++ FemaleDaily/FemaleDaily/View/Map/Components/BrandCard.swift
