+++ FemaleDaily/FemaleDaily/View/Map/StaticMapView.swift
