+++ FemaleDaily/FemaleDaily/View/Queue/BrandQueueView.swift
@@ -10,92 +10,6 @@ import AuthenticationServices
 import CloudKit
 import UserNotifications
 
-class LoginViewModel: ObservableObject {
-    static let shared = LoginViewModel()
-    
-    @Published var isLoggedIn = false
-    @Published var userName: String?
-    @Published var errorMessage: String?
-    @Published var isAdmin = false
-    
-    private init() {}
-    
-    func login(email: String, password: String) {
-        let emailLowercased = email.lowercased()
-        if emailLowercased == password.lowercased(), (emailLowercased.hasPrefix("user") || emailLowercased.hasPrefix("admin")) {
-            self.userName = emailLowercased
-            self.isAdmin = emailLowercased.hasPrefix("admin")
-            self.isLoggedIn = true
-            self.errorMessage = nil
-            print("Login success: \(emailLowercased), isAdmin: \(self.isAdmin)")
-        } else {
-            self.errorMessage = "Invalid credentials. Username must start with 'user' or 'admin' and match password."
-            self.isLoggedIn = false
-            self.isAdmin = false
-            print("Login failed for: \(emailLowercased)")
-        }
-    }}
-
-
-struct LoginView: View {
-    @State private var email = ""
-    @State private var password = ""
-    @Environment(\.colorScheme) var colorScheme
-    @State private var isLoggedIn = false
-
-    var body: some View {
-        NavigationStack {
-            VStack(spacing: 20) {
-                Text("Login to JxB")
-                    .font(.title)
-                    .fontWeight(.bold)
-                
-                if let errorMessage = LoginViewModel.shared.errorMessage {
-                    Text(errorMessage)
-                        .foregroundColor(.red)
-                        .font(.footnote)
-                        .multilineTextAlignment(.center)
-                }
-                
-                TextField("Username", text: $email)
-                    .textFieldStyle(.roundedBorder)
-                    .autocapitalization(.none)
-                
-                SecureField("Password", text: $password)
-                    .textFieldStyle(.roundedBorder)
-                
-                Button("Login") {
-                    LoginViewModel.shared.login(email: email, password: password)
-                    print("Login button pressed, isLoggedIn: \(LoginViewModel.shared.isLoggedIn), isAdmin: \(LoginViewModel.shared.isAdmin)")
-                    isLoggedIn = LoginViewModel.shared.isLoggedIn
-                    print("Set isLoggedIn to \(isLoggedIn)")
-                }
-                .frame(maxWidth: .infinity)
-                .padding()
-                .background(Color.blue)
-                .foregroundColor(.white)
-                .cornerRadius(8)
-            }
-            .padding()
-            .navigationDestination(isPresented: $isLoggedIn) {
-                if LoginViewModel.shared.isAdmin {
-                    AdminView()
-                        .environmentObject(QueueService.shared)
-                        .onAppear {
-                            print("Navigated to AdminView")
-                        }
-                } else {
-                    FlashSaleView()
-                        .environmentObject(QueueService.shared)
-                        .onAppear {
-                            print("Navigated to FlashSaleView")
-                        }
-                }
-            }
-        }
-    }
-}
-
 struct BrandQueueView: View {
     @EnvironmentObject private var queueService: QueueService
     @State private var isLoading = false
@@ -246,65 +160,3 @@ struct BrandQueueView: View {
         return queueService.queueList.contains { $0.username == username && $0.status.lowercased() == "queueing" }
     }
 }
-
-//// Home View
-//struct HomeView: View {
-//    @EnvironmentObject private var authViewModel: LoginViewModel
-//
-//    var body: some View {
-//        NavigationStack {
-//            VStack(spacing: 20) {
-//                Text("Welcome to JxB, \(authViewModel.userName ?? "User")!")
-//                    .font(.title2)
-//                NavigationLink("Go to Skintific Queue", destination: BrandQueueView(authViewModel: authViewModel))
-//                    .padding()
-//                    .background(Color.blue)
-//                    .foregroundColor(.white)
-//                    .cornerRadius(8)
-//            }
-//            .navigationTitle("Home")
-//        }
-//    }
-//}
-
-// AppDelegate untuk notifikasi CloudKit
-//class AppDelegate: NSObject, UIApplicationDelegate {
-//    
-//    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-//        UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .sound]) { granted, error in
-//            if let error = error {
-//                print("Notification authorization failed: \(error.localizedDescription)")
-//            } else {
-//                print("Notification authorization granted: \(granted)")
-//            }
-//        }
-//        UIApplication.shared.registerForRemoteNotifications()
-//        return true
-//    }
-//    
-//    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-//        print("Registered for push notifications with token: \(deviceToken.map { String(format: "%02.2hhx", $0) }.joined())")
-//    }
-//    
-//    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
-//        print("Failed to register for push notifications: \(error.localizedDescription)")
-//    }
-//    
-//    func application(_ application: UIApplication, didReceiveRemoteNotification userInfo: [AnyHashable: Any], fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
-//        print("Received notification at \(Date().formatted()) with userInfo: \(userInfo)")
-//        if let notification = CKNotification(fromRemoteNotificationDictionary: userInfo) {
-//            print("Received CloudKit notification at \(Date().formatted()): \(notification)")
-//            NotificationCenter.default.post(name: NSNotification.Name("QueueUpdated"), object: nil)
-//            completionHandler(.newData)
-//        } else {
-//            print("Received non-CloudKit notification at \(Date().formatted())")
-//            completionHandler(.noData)
-//        }
-//    }
-//}
-
-//
-//#Preview {
-//    LoginView()
-//        .environmentObject(LoginViewModel())
-//}
